Add ContainerAllocator

+++ ServerEngine/Core/Tls.cpp
@@ -2,5 +2,5 @@
 
 #include "ServerEngine/Pch.h"
 
-thread_local Int16 tThreadId = 0;
-thread_local Stack<Int32> tLockStack;
+thread_local Int32          tThreadId = 0;
+thread_local Stack<Int32>   tLockStack;
