+++ ServerEngine/Core/Tls.h
@@ -2,5 +2,5 @@
 
 #pragma once
 
-extern thread_local Int16 tThreadId;
-extern thread_local Stack<Int32> tLockStack;
+extern thread_local Int32           tThreadId;
+extern thread_local Stack<Int32>    tLockStack;
